Allow temp dir name override
Allow for the temp directory the Image Tool uses to be overridden and specified by the user using an environment variable. This makes it possible to pin the directory used, making it now possible to pin all the directories the Image Tool uses to specific paths.

As this directory is also used as the Docker build context, this also allows for preparing additional files in advance to be included and thus be available during the different additionalBuildCommands stages. This would be a solution for #106 .